update logo and favicon, use navmenu and footer from ds 1.0.7
