+++ package.json
@@ -19,6 +19,7 @@
     "@bitget-wallet/aptos-wallet-adapter": "^0.1.2",
     "@blocto/aptos-wallet-adapter-plugin": "^0.2.11",
     "@download/blockies": "^1.0.3",
+    "@moveindustries/movement-design-system": "1.0.7",
     "@martianwallet/aptos-wallet-adapter": "^0.0.5",
     "@msafe/aptos-wallet-adapter": "^1.1.8",
     "@mui/icons-material": "^6.4.3",
